refactor(design-system): move ToggleButtonGroup code
