+++ packages/design-system/src/components/ToggleButtonGroup/ToggleButtonGroup.tsx
@@ -1,0 +1,13 @@
+import React from "react";
+import MuiToggleButtonGroup, {
+  ToggleButtonGroupProps,
+} from "@mui/material/ToggleButtonGroup";
+
+const ToggleButtonGroup = ({
+  size = "small",
+  ...props
+}: ToggleButtonGroupProps) => {
+  return <MuiToggleButtonGroup size={size} {...props} />;
+};
+
+export default ToggleButtonGroup;
