+++ packages/design-system/src/components/ToggleButtonGroup/index.tsx
@@ -1,13 +1,1 @@
-import React from "react";
-import MuiToggleButtonGroup, {
-  ToggleButtonGroupProps,
-} from "@mui/material/ToggleButtonGroup";
-
-const ToggleButtonGroup = ({
-  size = "small",
-  ...props
-}: ToggleButtonGroupProps) => {
-  return <MuiToggleButtonGroup size={size} {...props} />;
-};
-
-export default ToggleButtonGroup;
+export { default } from "./ToggleButtonGroup";
